Smaller fn for selecting current top level form
This is still a slightly hacky way of doing this -- it assumes the top level open paren of a form is on the first column of a line, so it could send big sections if, for example, you're wrapping a bunch of stuff in a macro form.

But, IMO, it's better than relying on blank lines.  I rely on newlines for clarity within forms, so that way's a dealbreaker for me.