Attempt to fix borked undo
This needs the latest Vim (8.1) with patch 245.

The idea is to invoke parinfer only during regular edits but not after
an undo since invoking parinfer after an undo creates a branch in the
undo tree causing previous changes to not be easily reachable and the
user to be caught in an undo loop.