Use less conflicting mappings
Use `>>` and `<<` for indenting and dedenting instead of `<Tab>` and `<S-Tab>`.
Mapping `<Tab>` in normal mode also overrides `<C-i>`, which is used for
jumping forward through the jumplist.

As far as I can tell, these mappings are only needed when the version
of Vim does not support the `TextChanged` autocommand event. The
mappings will no longer be defined otherwise.

This change also includes a level of indirection by introducing the
`<Plug>Parinfer*` mappings. These allow a user to more easily define
their own mappings, i.e.:

    nmap <Tab> <Plug>ParinferDoIndent
    nmap <S-Tab> <Plug>ParinferDoUndent

This also allows the `do_indent` and `do_undent` functions to be
made private (not in this change).